Rename DebugData.js -> LocationData file
