--- app/components/AddressResult/index.js
@@ -1,4 +1,0 @@
-import AddressResult from './AddressResult';
-import DebugData from './DebugData';
-
-export { AddressResult, DebugData };
