+++ app/components/LocationData/index.js
@@ -1,0 +1,4 @@
+import LocationResult from './LocationResult';
+import LocationData from './LocationData';
+
+export { LocationResult, LocationData };
